ZainIslamicBot/requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
pytz==2023.3
hijri-converter==2.3.0